fix:fix bugs

+++ src/controllers/auditionpost.controllers.js
@@ -18,12 +18,20 @@ async function httpCreateAuditionPost(req,res){
 }
 
 async function httpGetAuditionPosts(req,res){
+    const auditionsExist = await getAuditionPosts();
+    if (!auditionsExist){
+        throw new ErrorResponse('There is no audition to show',404);
+    }
     return res
             .status(200)
             .json(await getAuditionPosts());
 }
 
 async function httpGetAuditionPost(req,res){
+    const auditionExists = await getAuditionPost({_id: req.params.id});
+    if (!auditionExists){
+        throw new ErrorResponse('Audition does not exist',404);
+    }
     return res
             .status(200)
             .json(await getAuditionPost(req.params.id));
@@ -31,19 +39,33 @@ async function httpGetAuditionPost(req,res){
 
 async function httpUpdateAuditionPost(req,res){
     const auditionpost = await getAuditionPost(req.params.id);
-    if (auditionpost.author.toString() !== req.user._id.toString()) throw new ErrorResponse("You're not authorized to do this", 403);
+    if (!auditionpost){
+        throw new ErrorResponse('Audition does not exist',404);
+    }
+    if (auditionpost.author._id.toString() !== req.user._id.toString()) throw new ErrorResponse("You're not authorized to do this", 403);
+
+       
+
     return res
             .status(200)
             .json(await updateAuditionPost(auditionpost.id,req.body));
 }
 
 async function httpDeleteAuditionPost(req,res) {
+    const auditionExists = await getAuditionPost({_id: req.params.id});
+    if (!auditionExists){
+        throw new ErrorResponse('Audition does not exist',404);
+    }
     return res
             .status(200)
             .json(await deleteAuditionPost(req.params.id));
 }
 
 async function httpDeleteAuditionPosts(req,res) {
+    const auditionsExist = await getAuditionPosts();
+    if (!auditionsExist){
+        throw new ErrorResponse('There is no audition to show',404);
+    }
     return res
             .status(200)
             .json(await deleteAuditionPosts());
